backend/requirements.txt
fastapi
uvicorn
pandas
numpy
python-docx
pdfplumber
PyMuPDF
spacy
fuzzywuzzy
sentence-transformers
openai
sqlalchemy
pydantic
python-dotenv